requirements.txt
beautifulsoup4==4.12.3
certifi==2024.6.2
charset-normalizer==3.3.2
greenlet==3.0.3
idna==3.7
playwright==1.44.0
pyee==11.1.0
python-dotenv==1.0.1
requests==2.32.3
ruff==0.4.8
soupsieve==2.5
typing-extensions==4.12.2
urllib3==2.2.1
